admin: Move the Figma CLI's exit-code mapping to its error module
The mapping covered only network errors when it was written, so it
lived in the HTTP client. It now also covers `component_unknown` and
`node_missing`, which the client never raises.
